Se crea un nueva funcion o api que regresa una lista de productos en base al negocio que se manda

+++ src/main/java/hn/unah/lenguajes1900/delivery/delivery/Services/ProductoService.java
@@ -9,6 +9,6 @@ public interface ProductoService {
 
     public Boolean CrearProducto(Producto producto);
 
-    public List<Producto> TraerProductosxNegocio(Float idNegocio);
+    public List<Producto> TraerProductosxNegocio(Long idNegocio);
 
 }
